Revert "Fixed issue where cactuses destroyed the universe. Added character selection."
This reverts commit 2cb2a7542b78583b4b2068a6a03ce65d09954ae9.
