--- Assets/Objects/Players/PlayersWithCameras.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5cc088d975553db48bb3fce3e4fa185d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
