--- Assets/Objects/Players/PlayersWithCameras/Cactus.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 75a7a70cd26e5c44a985a8261d8af668
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
