--- Assets/Objects/Players/PlayersWithCameras/Mushroom.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c313af6c7c1a4194097d50658ac95fd1
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
