--- Assets/Objects/Players/PlayersWithCameras/Penguin.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4f0d5ea59fa24114cba0e9500f272674
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
