--- Assets/Objects/Players/PlayersWithCameras/Pig.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d6db2742dbcd1b448b5057f7c97cbd4e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
