--- Assets/Objects/Players/PlayersWithCameras/Plant.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0b4fbe3d434d2744383388a90e339ae0
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
