+++ Assets/Scripts/PlayerChooser.cs
@@ -19,7 +19,6 @@ public class PlayerChooser : MonoBehaviour
         PlayerPrefs.SetFloat("ClimbSpeed", 0.5f);
         PlayerPrefs.SetFloat("SwimSpeed", 0.5f);
         PlayerPrefs.SetFloat("FlightSpeed", 0f);
-        PlayerPrefs.SetString("Character","Jumper");
         LoadLevel();
     }
     public void SetRunner()
@@ -29,7 +28,6 @@ public class PlayerChooser : MonoBehaviour
         PlayerPrefs.SetFloat("ClimbSpeed", 0.5f);
         PlayerPrefs.SetFloat("SwimSpeed", 0.5f);
         PlayerPrefs.SetFloat("FlightSpeed", 0f);
-        PlayerPrefs.SetString("Character","Runner");
         LoadLevel();
     }
     public void SetClimber()
@@ -39,7 +37,6 @@ public class PlayerChooser : MonoBehaviour
         PlayerPrefs.SetFloat("ClimbSpeed", 1.0f);
         PlayerPrefs.SetFloat("SwimSpeed", 0.5f);
         PlayerPrefs.SetFloat("FlightSpeed", 0f);
-        PlayerPrefs.SetString("Character","Climber");
         LoadLevel();
     }
     public void SetFlyer()
@@ -49,7 +46,6 @@ public class PlayerChooser : MonoBehaviour
         PlayerPrefs.SetFloat("ClimbSpeed", 0.5f);
         PlayerPrefs.SetFloat("SwimSpeed", 0.5f);
         PlayerPrefs.SetFloat("FlightSpeed", 1.0f);
-        PlayerPrefs.SetString("Character","Flyer");
         LoadLevel();
     }
     public void SetSwimmer()
@@ -59,7 +55,6 @@ public class PlayerChooser : MonoBehaviour
         PlayerPrefs.SetFloat("ClimbSpeed", 0.5f);
         PlayerPrefs.SetFloat("SwimSpeed", 2.0f);
         PlayerPrefs.SetFloat("FlightSpeed", 0f);
-        PlayerPrefs.SetString("Character","Swimmer");
         LoadLevel();
     }
 }
