+++ Assets/Scripts/PlayerMovement.cs
@@ -105,7 +105,7 @@ public float ForwardsForce;
         {
             rb.AddForce(0, JumpForce * Time.deltaTime, 0);
         }
-        if(rb.position.y < -5)
+        if(rb.position.y < .8)
         {
             FindObjectOfType<GameManager>().EndGame();
         }
